moving ReservoirSampler and UniformSampler to their own file
